+++ scripts/mhs-pkg.sh
@@ -9,4 +9,4 @@ MHS_MIDI=`pwd`/projects/mhs-midi
 
 MHSDIR=${THIRDPARTY}/MicroHs ${THIRDPARTY}/MicroHs/bin/mhs -P${PKG_NAME}-${PKG_VER} -i${MHS_MIDI}/lib \
 	Async Midi MidiPerform Music MusicPerform \
-	-o ${PKG_NAME}.pkg
+	-o ${PKG_NAME}-${PKG_VER}.pkg